fix: nats retry logic
The access token is now fetched by the Nats client.
This ensures that a valid token is available for a reconnect.

A reconnect_delay_callback has also been added to avoid
constantly fetching new tokens, which increases the
waiting time for repeated attempts.

The provider now re-registers with every Nats connect,
as the registry removes it when the connection is lost.

A state machine for the worker thread has been added
so that the individual states are more clearly delineated.

Refs: #UC20-13730

Signed-off-by: Etienne Schmidt <etienne.schmidt@weidmueller.com>